Fix semaphore for images that are ready to present
Fixed thanks to [this discussion](https://github.com/Overv/VulkanTutorial/issues/407).
Basically, the swapchain image is not released until a new one is acquired
by the `vkQueuePresent`. As such, the "image ready for present" semaphore is not released either.
And somewhere along those lines, the validation layer is lost.

To fix it, we create a "image ready for present" semaphore for every swapchain image.
That way `vkQueuePresent` will not kidnap it until it gets a new image.

The "image available" semaphore does not need it, as it is released by `vkAcquireNextImageKHR`.